User Video finish

+++ src/controllers/videoController.js
@@ -4,6 +4,7 @@ import User from "../models/User";
 export const see = async (req, res) => {
   const { id } = req.params;
   const video = await Video.findById(id).populate("owner");
+
   // const owner = await User.findById(video.owner);
   if (!video) {
     return res.status(404).render("404");
@@ -14,19 +15,31 @@ export const see = async (req, res) => {
   });
 };
 export const getEdit = async (req, res) => {
-  const { id } = req.params;
+  const {
+    session: { _id },
+    params: { id },
+  } = req;
   const video = await Video.findById(id);
   if (!video) {
     return res.status(404).render("404");
   }
+  if (String(video.owner !== String(_id))) {
+    return res.status(403).redirect("/");
+  }
   return res.render("edit", { pageTitle: `Editing: ${video.title}`, video });
 };
 export const postEdit = async (req, res) => {
-  const { id } = req.params;
-  const { title, description, hashtags } = req.body;
+  const {
+    session: { _id },
+    params: { id },
+    body: { title, description, hashtags },
+  } = req;
   const video = await Video.exists({ _id: id });
   if (!video) {
     return res.status(404).render("404");
+  }
+  if (String(video.owner !== String(_id))) {
+    return res.status(403).redirect("/");
   }
   await Video.findByIdAndUpdate(id, {
     title,
@@ -48,13 +61,18 @@ export const postUpload = async (req, res) => {
   } = req;
 
   try {
-    await Video.create({
+    const newVideo = await Video.create({
       title,
       description,
       fileUrl: path,
       owner: _id,
       hashtags: Video.formatHashtags(hashtags),
     });
+    const user = await User.findById(_id);
+
+    user.videos.push(newVideo._id);
+    user.save();
+
     res.redirect("/");
   } catch (error) {
     res.status(400).render("upload", {
@@ -65,7 +83,17 @@ export const postUpload = async (req, res) => {
 };
 
 export const deleteVideo = async (req, res) => {
-  const { id, createdAt } = req.params;
+  const {
+    session: { _id },
+    params: { id },
+  } = req;
+  const video = await Video.findById(id);
+  if (!video) {
+    return res.status(404).render("404", { pageTitle: "Video not found." });
+  }
+  if (String(video.owner !== String(_id))) {
+    return res.status(403).redirect("/");
+  }
   await Video.findByIdAndDelete(id);
   return res.redirect("/");
 };
